--- server/datastore/mysql/datastore_test.go
@@ -1,12 +1,0 @@
-package mysql
-
-import (
-	"testing"
-
-	"github.com/fleetdm/fleet/v4/server/datastore"
-	_ "github.com/go-sql-driver/mysql"
-)
-
-func TestMySQL(t *testing.T) {
-	RunTestsAgainstMySQL(t, datastore.TestFunctions)
-}
