+++ server/datastore/mysql/mysql_test.go
@@ -1,6 +1,7 @@
 package mysql
 
 import (
+	"os"
 	"testing"
 
 	"github.com/DATA-DOG/go-sqlmock"
@@ -14,6 +15,14 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+func TestMain(m *testing.M) {
+	if _, ok := os.LookupEnv("MYSQL_TEST"); ok {
+		// Initialize the schema once for the entire test run.
+		initializeSchemaOrPanic()
+	}
+	os.Exit(m.Run())
+}
 
 func TestSanitizeColumn(t *testing.T) {
 	t.Parallel()
