+++ server/datastore/mysql/unicode_test.go
@@ -1,9 +1,10 @@
-package datastore
+package mysql
 
 import (
-	"github.com/fleetdm/fleet/v4/server/ptr"
 	"testing"
 	"time"
+
+	"github.com/fleetdm/fleet/v4/server/ptr"
 
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/fleetdm/fleet/v4/server/test"
@@ -11,7 +12,10 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func testUnicode(t *testing.T, ds fleet.Datastore) {
+func TestUnicode(t *testing.T) {
+	ds := CreateMySQLDS(t)
+	defer ds.Close()
+
 	if ds.Name() == "inmem" {
 		t.Skip("inmem is being deprecated, test skipped")
 	}
